Ref doctest repo to my fork with CMake uprev
Latest versions of CMake will deprecate compatibility with
versions < 3.5. Doctest's CMakeLists has min 3.0, causing
warnings when configuring doctest.

The solution is to update doctest upstream to uprev this,
but while that ticket waits, I'll patch in a personal fork
and point this to that repo. See gunvirranu/doctest for diff.